Java Program to check in array index 3 is having Vowel or not.
In this Program we make array of 5 names and that 5 names enter by user  after user enter the name it store in array.
After this we separate the  character which is present at index 3 in given names and check that particular character is vowel or not .Using charAt(index no.) method  we seperate the charater which is present at index 3 in given name.
If that  index 3 character is vowel then print message
In given name at index 3 vowel is present,Vowel is : Print the given vowel character present at index 3

Output look like :
In Ambar at index 3 vowel is present,Vowel is : a